+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.4.0
+
+## Prefix for inputs
+
+We have added a new feature for input components which allows developers to output a prefix to the input.
+
+```js
+<Textbox prefix="foo" />
+```
+
 # 0.3.3
 
 * Performance updates to inputs. We also now provide a `shouldComponentUpdate` method which can be reused in custom components.